Add rules' refreshing after deletion of a group

+++ public/app/features/alerting/unified/state/actions.ts
@@ -277,7 +277,15 @@ export function deleteRulesGroupAction(
     withAppEvents(
       (async () => {
         const sourceName = getRulesSourceName(namespace.rulesSource);
+
+        const freshGroup = fetchRulerRulesGroup(sourceName, namespace.name, ruleGroup.name);
+        if (!freshGroup) {
+          throw new Error('Rule group not found.');
+        }
+
         await deleteRulerRulesGroup(sourceName, namespace.name, ruleGroup.name);
+        dispatch(fetchRulerRulesAction({ rulesSourceName: sourceName }));
+        dispatch(fetchPromRulesAction({ rulesSourceName: sourceName }));
       })(),
       { successMessage: 'Group deleted' }
     );
